+++ source/B2B.Transactions.Infrastructure/IncomingMessages/MessageIdRegistry.cs
@@ -18,25 +18,25 @@ using B2B.CimMessageAdapter.Messages;
 using B2B.Transactions.Configuration.DataAccess;
 using Dapper;
 
-namespace B2B.Transactions.Infrastructure.Messages
+namespace B2B.Transactions.Infrastructure.IncomingMessages
 {
-    public class TransactionIdRegistry : ITransactionIds
+    public class MessageIdRegistry : IMessageIds
     {
         private readonly IDbConnectionFactory _connectionFactory;
 
-        public TransactionIdRegistry(IDbConnectionFactory connectionFactory)
+        public MessageIdRegistry(IDbConnectionFactory connectionFactory)
         {
             _connectionFactory = connectionFactory ?? throw new ArgumentNullException(nameof(connectionFactory));
         }
 
-        public async Task<bool> TryStoreAsync(string transactionId)
+        public async Task<bool> TryStoreAsync(string messageId)
         {
             var connection = _connectionFactory.GetOpenConnection();
 
             var result = await connection.ExecuteAsync(
-                    $"IF NOT EXISTS (SELECT * FROM b2b.TransactionIds WHERE TransactionId = @TransactionId)" +
-                    $"INSERT INTO b2b.TransactionIds(TransactionId) VALUES(@TransactionId)",
-                    new { TransactionId = transactionId })
+                    $"IF NOT EXISTS (SELECT * FROM b2b.MessageIds WHERE MessageId = @MessageId)" +
+                    $"INSERT INTO b2b.MessageIds(MessageId) VALUES(@MessageId)",
+                    new { MessageId = messageId })
                 .ConfigureAwait(false);
 
             return result == 1;
