+++ source/B2B.Transactions.Infrastructure/IncomingMessages/TransactionIdRegistry.cs
@@ -18,25 +18,25 @@ using B2B.CimMessageAdapter.Messages;
 using B2B.Transactions.Configuration.DataAccess;
 using Dapper;
 
-namespace B2B.Transactions.Infrastructure.Messages
+namespace B2B.Transactions.Infrastructure.IncomingMessages
 {
-    public class MessageIdRegistry : IMessageIds
+    public class TransactionIdRegistry : ITransactionIds
     {
         private readonly IDbConnectionFactory _connectionFactory;
 
-        public MessageIdRegistry(IDbConnectionFactory connectionFactory)
+        public TransactionIdRegistry(IDbConnectionFactory connectionFactory)
         {
             _connectionFactory = connectionFactory ?? throw new ArgumentNullException(nameof(connectionFactory));
         }
 
-        public async Task<bool> TryStoreAsync(string messageId)
+        public async Task<bool> TryStoreAsync(string transactionId)
         {
             var connection = _connectionFactory.GetOpenConnection();
 
             var result = await connection.ExecuteAsync(
-                    $"IF NOT EXISTS (SELECT * FROM b2b.MessageIds WHERE MessageId = @MessageId)" +
-                    $"INSERT INTO b2b.MessageIds(MessageId) VALUES(@MessageId)",
-                    new { MessageId = messageId })
+                    $"IF NOT EXISTS (SELECT * FROM b2b.TransactionIds WHERE TransactionId = @TransactionId)" +
+                    $"INSERT INTO b2b.TransactionIds(TransactionId) VALUES(@TransactionId)",
+                    new { TransactionId = transactionId })
                 .ConfigureAwait(false);
 
             return result == 1;
